.member-auth분리,role삭제,인증/인가 config파일 위치 변경
