+++ src/main/java/com/sparta/spartanewsfeed/domain/jwt/jwt/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.sparta.spartanewsfeed.domain.member.jwt;
+package com.sparta.spartanewsfeed.domain.jwt.jwt;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
@@ -12,8 +12,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
-
-import com.sparta.spartanewsfeed.domain.member.UserRole;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -59,13 +57,13 @@ public class JwtUtil {
 
 	//JWT 생성
 	// 토큰 생성
-	public String createToken(String username, UserRole role) {
+	public String createToken(String username) {
 		Date date = new Date();
 
 		return BEARER_PREFIX +
 			Jwts.builder()
 				.setSubject(username) // 사용자 식별자값(ID)
-				.claim(AUTHORIZATION_KEY, role) // 사용자 권한
+				//.claim(AUTHORIZATION_KEY, role) // 사용자 권한
 				.setExpiration(new Date(date.getTime() + TOKEN_TIME)) // 만료 시간
 				.setIssuedAt(date) // 발급일
 				.signWith(key, signatureAlgorithm) // 암호화 알고리즘
