+++ src/main/java/com/sparta/spartanewsfeed/domain/member/repository/AuthRepository.java
@@ -1,9 +1,11 @@
 package com.sparta.spartanewsfeed.domain.member.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sparta.spartanewsfeed.domain.member.Member;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
+public interface AuthRepository extends JpaRepository<Member, Long> {
+	Optional<Member> findByEmail(String email);
 }
